@@ -1,25 +1,32 @@
 {
     "name": "gizemsever/laravel-paytr",
-    "license": "MIT",
-    "type": "library",
     "description": "Laravel Paytr Wrapper",
+    "type": "library",
+    "license": "MIT",
     "authors": [
         {
             "name": "Gizem Sever",
             "email": "gizemsever68@gmail.com"
         }
     ],
-    "minimum-stability": "dev",
     "require": {
-        "php": "^8.0|^8.1",
+        "php": "^8.2",
         "guzzlehttp/guzzle": "^7.0.1",
-        "illuminate/collections": "^8|^9|^10",
-        "illuminate/support": "^8.0|^9.0|^10.0"
+        "illuminate/collections": "^8|^9|^10|^11|^12",
+        "illuminate/support": "^8.0|^9.0|^10.0|^11.0|^12.0"
     },
+    "require-dev": {
+        "laravel/pint": "dev-main"
+    },
+    "minimum-stability": "dev",
     "autoload": {
         "psr-4": {
             "Gizemsever\\LaravelPaytr\\": "src/"
         }
+    },
+    "scripts": {
+        "format": "vendor/bin/pint",
+        "format:test": "vendor/bin/pint --test"
     },
     "config": {
         "sort-packages": true
