@@ -1,4 +1,5 @@
 <?php
+
 /**
  * @author Gizem Sever <gizemsever68@gmail.com>
  */
@@ -11,17 +12,27 @@ use Illuminate\Http\Response;
 
 class PaymentVerification extends PaytrClient
 {
-    private $request;
-    private $merchantOid;
-    private $status;
-    private $totalAmount;
-    private $hash;
-    private $failedReasonCode;
-    private $failedReasonMessage;
-    private $testMode;
-    private $paymentType;
-    private $currency;
-    private $paymentAmount;
+    private Request $request;
+
+    private string $merchantOid;
+
+    private string $status;
+
+    private float $totalAmount;
+
+    private string $hash;
+
+    private ?int $failedReasonCode;
+
+    private ?string $failedReasonMessage;
+
+    private ?int $testMode;
+
+    private ?string $paymentType;
+
+    private ?string $currency;
+
+    private mixed $paymentAmount;
 
     public function __construct(Request $request)
     {
@@ -29,7 +40,7 @@ class PaymentVerification extends PaytrClient
         $this->setup();
     }
 
-    private function setup()
+    private function setup(): void
     {
         $this->merchantOid = $this->request->input('merchant_oid');
         $this->status = $this->request->input('status');
@@ -43,101 +54,76 @@ class PaymentVerification extends PaytrClient
         $this->paymentAmount = $this->request->input('payment_amount');
     }
 
-    /**
-     * @return mixed
-     */
-    public function getMerchantOid()
+    public function getMerchantOid(): string
     {
         return $this->merchantOid;
     }
 
-    /**
-     * @return mixed
-     */
-    public function getTotalAmount()
+    public function getTotalAmount(): float
     {
         return $this->totalAmount;
     }
 
-    /**
-     * @return mixed
-     */
-    public function getHash()
+    public function getHash(): string
     {
         return $this->hash;
     }
 
-    /**
-     * @return mixed
-     */
-    public function getFailedReasonCode()
+    public function getFailedReasonCode(): ?int
     {
         return $this->failedReasonCode;
     }
 
-    /**
-     * @return mixed
-     */
-    public function getFailedReasonMessage()
+    public function getFailedReasonMessage(): ?string
     {
         return $this->failedReasonMessage;
     }
 
-    /**
-     * @return mixed
-     */
-    public function getTestMode()
+    public function getTestMode(): ?int
     {
         return $this->testMode;
     }
 
-    /**
-     * @return mixed
-     */
-    public function getPaymentType()
+    public function getPaymentType(): ?string
     {
         return $this->paymentType;
     }
 
-    /**
-     * @return mixed
-     */
-    public function getCurrency()
+    public function getCurrency(): ?string
     {
         return $this->currency;
     }
 
-    /**
-     * @return mixed
-     */
-    public function getPaymentAmount()
+    public function getPaymentAmount(): mixed
     {
         return $this->paymentAmount;
     }
 
-    public function verifyRequest()
+    public function verifyRequest(): bool
     {
         $generatedHash = $this->generateHash();
-        return ($generatedHash === $this->hash);
+
+        return $generatedHash === $this->hash;
     }
 
-    public function isSuccess()
+    public function isSuccess(): bool
     {
-        return ($this->status === 'success');
+        return $this->status === 'success';
     }
 
-    public function getProcessedResponse()
+    public function getProcessedResponse(): Response
     {
         return response('OK', Response::HTTP_OK);
     }
 
-    private function generateHash()
+    private function generateHash(): string
     {
-        $hashStr = '' .
-            $this->merchantOid .
-            $this->credentials['merchant_salt'] .
-            $this->status .
+        $hashStr = ''.
+            $this->merchantOid.
+            $this->credentials['merchant_salt'].
+            $this->status.
             $this->totalAmount;
+
         return base64_encode(hash_hmac('sha256', $hashStr, $this->credentials['merchant_key'], true));
     }
 }
