@@ -1,4 +1,5 @@
 <?php
+
 /**
  * @author Gizem Sever <gizemsever68@gmail.com>
  */
@@ -6,31 +7,34 @@
 namespace Gizemsever\LaravelPaytr;
 
 use GuzzleHttp\ClientInterface;
+use Psr\Http\Message\ResponseInterface;
 
 class PaytrClient
 {
-    /**
-     * @var ClientInterface
-     */
-    protected $client;
-    protected $credentials = [];
-    protected $options = [];
+    protected ClientInterface $client;
 
-    public function setClient(ClientInterface $client)
+    protected array $credentials = [];
+
+    protected array $options = [];
+
+    public function setClient(ClientInterface $client): static
     {
         $this->client = $client;
+
         return $this;
     }
 
-    public function setCredentials(array $credentials)
+    public function setCredentials(array $credentials): static
     {
         $this->credentials = $credentials;
+
         return $this;
     }
 
-    public function setOptions(array $options)
+    public function setOptions(array $options): static
     {
         $this->options = $options;
+
         return $this;
     }
 
@@ -39,22 +43,26 @@ class PaytrClient
         return base64_encode(hash_hmac('sha256', $hash, $this->credentials['merchant_key'], true));
     }
 
-    protected function callApi(string $method, string $url, $params = null, $headers = null)
+    protected function callApi(string $method, string $url, ?array $params = null, ?array $headers = null): ResponseInterface
     {
         $options = [];
+
         if ($headers) {
             $options['headers'] = $headers;
         }
+
         if ($params) {
             $requestParams = collect(array_keys($params));
             $options['multipart'] = $requestParams->map(function ($key) use ($params) {
                 return [
                     'name' => $key,
-                    'contents' => $params[$key]
+                    'contents' => $params[$key],
                 ];
             })->toArray();
         }
+
         $options['timeout'] = $this->options['timeout'];
-        return $this->client->request($method, $this->options['base_uri'] . '/' . $url, $options);
+
+        return $this->client->request($method, $this->options['base_uri'].'/'.$url, $options);
     }
 }
