@@ -1,4 +1,5 @@
 <?php
+
 /**
  * @author Gizem Sever <gizemsever68@gmail.com>
  */
@@ -10,44 +11,28 @@ use Gizemsever\LaravelPaytr\PaytrResponse;
 
 class BankIdentification extends PaytrClient
 {
-
-    /**
-     * @var string|null
-     */
     private ?string $binNumber;
 
-    /**
-     * @return string|null
-     */
     public function getBinNumber(): ?string
     {
         return $this->binNumber;
     }
 
-    /**
-     * @param string|null $binNumber
-     * @return BankIdentification
-     */
     public function setBinNumber(?string $binNumber): static
     {
         $this->binNumber = $binNumber;
+
         return $this;
     }
 
-    /**
-     * @return string
-     */
     private function getHash(): string
     {
-        return '' .
-            $this->getBinNumber() .
-            $this->credentials['merchant_id'] .
+        return ''.
+            $this->getBinNumber().
+            $this->credentials['merchant_id'].
             $this->credentials['merchant_salt'];
     }
 
-    /**
-     * @return PaytrResponse
-     */
     public function checkBin(): PaytrResponse
     {
         $hash = $this->getHash();
@@ -57,9 +42,9 @@ class BankIdentification extends PaytrClient
             'bin_number' => $this->getBinNumber(),
             'paytr_token' => $token,
         ];
+
         $response = $this->callApi('POST', 'odeme/api/bin-detail', $body);
 
-        return new PaytrResponse(json_decode((string)$response->getBody(), true));
+        return new PaytrResponse(json_decode((string) $response->getBody(), true));
     }
-
 }
