@@ -1,4 +1,5 @@
 <?php
+
 /**
  * @author Gizem Sever <gizemsever68@gmail.com>
  */
@@ -12,27 +13,21 @@ class Capi extends PaytrClient
 {
     private ?string $uToken;
 
-    /**
-     * @return string|null
-     */
     public function getUToken(): ?string
     {
         return $this->uToken;
     }
 
-    /**
-     * @param string|null $uToken
-     * @return Capi
-     */
     public function setUToken(?string $uToken): static
     {
         $this->uToken = $uToken;
+
         return $this;
     }
 
-    public function getList()
+    public function getList(): PaytrResponse
     {
-        $hash = $this->getUToken() . $this->credentials['merchant_salt'];
+        $hash = $this->getUToken().$this->credentials['merchant_salt'];
         $token = $this->generateToken($hash);
 
         $body = [
@@ -40,8 +35,9 @@ class Capi extends PaytrClient
             'utoken' => $this->getUToken(),
             'paytr_token' => $token,
         ];
+
         $response = $this->callApi('POST', 'odeme/capi/list', $body);
 
-        return new PaytrResponse(json_decode((string)$response->getBody(), true));
+        return new PaytrResponse(json_decode((string) $response->getBody(), true));
     }
 }
