@@ -1,4 +1,5 @@
 <?php
+
 /**
  * @author Gizem Sever <gizemsever68@gmail.com>
  */
@@ -7,7 +8,7 @@ namespace Gizemsever\LaravelPaytr\Exceptions;
 
 class InvalidConfigException extends PaytrException
 {
-    public static function configNotFound()
+    public static function configNotFound(): static
     {
         return new static('Setup your credentials to config.paytr');
     }
