@@ -1,4 +1,5 @@
 <?php
+
 /**
  * @author Gizem Sever <gizemsever68@gmail.com>
  */
@@ -6,23 +7,23 @@
 namespace Gizemsever\LaravelPaytr;
 
 use Gizemsever\LaravelPaytr\Exceptions\InvalidConfigException;
-use Gizemsever\LaravelPaytr\Payment\Payment;
 use GuzzleHttp\Client;
 use Illuminate\Support\ServiceProvider;
 
 class PaytrServiceProvider extends ServiceProvider
 {
-    public function boot()
+    public function boot(): void
     {
-        $configPath = __DIR__ . '/../config/paytr.php';
+        $configPath = __DIR__.'/../config/paytr.php';
         $this->publishes([$configPath => config_path('paytr.php')], 'config');
     }
 
-    public function register()
+    public function register(): void
     {
-        $this->mergeConfigFrom(__DIR__ . '/../config/paytr.php', 'paytr');
+        $this->mergeConfigFrom(__DIR__.'/../config/paytr.php', 'paytr');
         $this->app->singleton(Paytr::class, function ($app) {
             $config = config('paytr');
+
             if (is_null($config)) {
                 throw InvalidConfigException::configNotFound();
             }
@@ -31,6 +32,7 @@ class PaytrServiceProvider extends ServiceProvider
                 'base_uri' => $config['options']['base_uri'],
                 'timeout' => $config['options']['timeout'],
             ]);
+
             return new Paytr($client, $config['credentials'], $config['options']);
         });
     }
